pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "pybind11"]
build-backend = "setuptools.build_meta"

[project]
name = "text2terrain"
version = "0.1.0"
description = "Text-to-terrain generation using neural procedural synthesis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Text2Terrain Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core computation
    "jax>=0.4.20",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "numpy>=1.24.0",
    
    # Essential utilities
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    
    # Inference API (minimal)
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "jupyter>=1.0.0",
]

training = [
    "accelerate>=0.24.0",
    "datasets>=2.14.0", 
    "wandb>=0.16.0",
]

bindings = [
    "pybind11>=2.11.0",
]

rendering = [
    "raylib-python>=4.5.0",
]

[project.scripts]
t2t-generate = "src.data_gen.generator:main"
t2t-train = "src.training.train:main"
t2t-serve = "src.inference.api:main"
t2t-build = "scripts.build_all:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true